added confirmation

+++ clear-instance.sh
@@ -5,14 +5,19 @@ set -euxo pipefail
 CONFIG_DIR=`dirname "$0"`
 source ${CONFIG_DIR}/master-config.shinc
 
-rm -r -f ${BUILD_DIR}
-rm -r -f ${CODE_DIR}
-rm -r -f ${VENV_DIR}
-rm -r -f ~/RTX
+read -p "Are you sure you are running this command in the KG2 build instance and not on your laptop? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then   
+    rm -r -f ${BUILD_DIR}
+    rm -r -f ${CODE_DIR}
+    rm -r -f ${VENV_DIR}
+    rm -r -f ~/RTX
+    sudo apt-get purge -y mysql-server mysql-client mysql-common mysql-server-core-* mysql-client-core-*
+    sudo rm -rf /etc/mysql /var/lib/mysql
+    sudo apt-get -y autoremove
+    sudo apt-get -y autoclean
+    rm -r -f ~/.cachier
+    rm -r -f ~/*.log
+fi
 
-sudo apt-get purge -y mysql-server mysql-client mysql-common mysql-server-core-* mysql-client-core-*
-sudo rm -rf /etc/mysql /var/lib/mysql
-sudo apt-get -y autoremove
-sudo apt-get -y autoclean
-rm -r -f ~/.cachier
-rm -r -f ~/*.log
